+++ common/errcode/prefix.go
@@ -1,0 +1,8 @@
+package errcode
+
+// 业务码-错误码前缀
+const (
+	LoginErrorPrefix = 1000 // 登录错误码前缀
+	UserErrorPrefix  = 2000 // 用户错误码前缀
+	MenuErrorPrefix  = 3000 // 菜单错误码前缀
+)