+++ common/response/errcode.go
@@ -5,7 +5,6 @@ import (
 	"gin/config"
 	"gin/pkg/provider/logger"
 	"github.com/gin-gonic/gin"
-	"net/http"
 )
 
 // Response 通用响应结构体
@@ -16,73 +15,63 @@ type Response struct {
 }
 
 // Json 输出Json响应
-func (r Response) Json(c *gin.Context) {
+func (r Response) Json(c *gin.Context, httpCode int) {
 	c.Header("Content-Type", "application/json")
-	c.JSON(http.StatusOK, r)
+	c.JSON(httpCode, r)
 	c.Abort()
 }
 
 // Success 返回成功响应,可传ErrorCode
-func Success(c *gin.Context, e *errcode.ErrorCode) {
+func Success(c *gin.Context, e errcode.ErrorCode) {
 	var (
 		resp Response
 	)
 
-	switch e {
-
-	case nil:
-		resp = Response{
-			Code: errcode.Success().Code,
-			Msg:  errcode.Success().Msg,
-			Data: []string{},
-		}
-
-	default:
-		if e.Data == nil {
-			e.Data = []string{}
-		}
-		resp = Response{
-			Code: e.Code,
-			Msg:  e.Msg,
-			Data: e.Data,
-		}
+	if e.Msg == "" {
+		e.Msg = errcode.Success().Msg
+	}
+	if e.Data == nil {
+		e.Data = []string{}
+	}
+	if e.HttpCode == 0 {
+		e.HttpCode = 200
+	}
+	resp = Response{
+		Code: e.Code,
+		Msg:  e.Msg,
+		Data: e.Data,
 	}
 
-	resp.Json(c)
+	resp.Json(c, e.HttpCode)
 }
 
 // Error 返回失败响应,可传ErrorCode
-func Error(c *gin.Context, e *errcode.ErrorCode) {
+func Error(c *gin.Context, e errcode.ErrorCode) {
 	var (
 		resp Response
 		ctx  = c.Request.Context()
 	)
 
-	if e != nil {
+	if e.Msg != "" {
 		logger.NewLogger(config.NewConfig()).WithDebugger(ctx).Error(e.Msg)
 	} else {
 		logger.NewLogger(config.NewConfig()).WithDebugger(ctx).Error(errcode.SystemError().Msg)
 	}
 
-	switch e {
-
-	case nil:
-		resp = Response{
-			Code: errcode.SystemError().Code,
-			Msg:  errcode.SystemError().Msg,
-			Data: []string{},
-		}
-
-	default:
-		if e.Data == nil {
-			e.Data = []string{}
-		}
-		resp = Response{
-			Code: e.Code,
-			Msg:  e.Msg,
-			Data: e.Data,
-		}
+	if e.Msg == "" {
+		e.Msg = errcode.SystemError().Msg
+	}
+	if e.Data == nil {
+		e.Data = []string{}
+	}
+	if e.HttpCode == 0 {
+		e.HttpCode = 500
+	}
+	resp = Response{
+		Code: e.Code,
+		Msg:  e.Msg,
+		Data: e.Data,
 	}
 
-	resp.Json(c)
+	resp.Json(c, e.HttpCode)
 }